pyproject.toml
[project]
name = "hr-recruiter-multi-agent-ai-workflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=1.7.2",
    "openai>=1.95.1",
    "pydantic>=2.11.7",
    "pypdf>=5.7.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "resend>=2.11.0",
]
